Include error class in message text.
Airbrake doesn't display /notice/error/class anywhere in the UI.  This kind of
annoying and upon futher inspection we see that the ruby library manually
adds the error class to the message string.  https://github.com/airbrake/airbrake/blob/master/lib/airbrake/notice.rb#L131

This seems like a pretty useful thing so we do that here too.

It's possible that we should also use the fullyQualifiedName of the exception class,
so however that would change a value that airbrake uses in fingerprinting so I'm
reluctant to do that.