Switch from getPage to twisted.web.client.Agent.
Since Twisted 9.0 getPage usage has been discouraged in favor of
twisted.web.client.Agent a more flexible HTTP/1.1 client library.

Since Twisted 12.1 (the last verison to support Python 2.5) Agent has included
an HTTPConnectionPool that allows for the usage of persistent connections.

This is preferrable to connecting for each exception AirbrakeLogObserver
reports.

We also update the setup.py to declare install_requires of Twisted >= 12.1.

We also add a dependency on mock for unit testing the usage of
the Agent interface.