Include framework frames in traceback.
failure.getTracebackObject() only includes stack frames between where
the exception was raised and where it was caught.

This is different from failure.printTraceback() which by default includes
the the entire stack.  This change causes the entire stack to be included in
the backtrace, which will cause many more stack frames to be logged but
is consistent with other log observers that use failure.printTraceback().